+++ src/Data/Set/NonEmpty.hs
@@ -90,6 +90,8 @@ module Data.Set.NonEmpty (
 
   -- * Combine
   union,
+  unionSetLeft,
+  unionSetRight,
   unions,
   difference,
   (\\),
@@ -485,6 +487,22 @@ disjoint n1@(NESet x1 s1) n2@(NESet x2 s2) = case compare x1 x2 of
   -- k2 is not in n1
   GT -> toSet n1 `S.disjoint` s2
 {-# INLINE disjoint #-}
+
+-- | /O(m*log(n\/m + 1)), m <= n/. Union of a possibly-empty 'Set' and a
+-- non-empty set.
+--
+-- @since 0.3.6.0
+unionSetLeft :: Ord a => Set a -> NESet a -> NESet a
+unionSetLeft s n = withNonEmpty n (`union` n) s
+{-# INLINE unionSetLeft #-}
+
+-- | /O(m*log(n\/m + 1)), m <= n/. Union of a non-empty set and a
+-- possibly-empty 'Set'.
+--
+-- @since 0.3.6.0
+unionSetRight :: Ord a => NESet a -> Set a -> NESet a
+unionSetRight n s = withNonEmpty n (union n) s
+{-# INLINE unionSetRight #-}
 
 -- | /O(m*log(n\/m + 1)), m <= n/. Difference of two sets.
 --