+++ test/Tests/Map.hs
@@ -443,6 +443,20 @@ prop_union =
     M.union
     NEM.union
 
+prop_unionMapLeft :: Property
+prop_unionMapLeft =
+  ttProp
+    (GTMap :-> GTNEMap :-> TTNEMap)
+    M.union
+    NEM.unionMapLeft
+
+prop_unionMapRight :: Property
+prop_unionMapRight =
+  ttProp
+    (GTNEMap :-> GTMap :-> TTNEMap)
+    M.union
+    NEM.unionMapRight
+
 prop_unionWith :: Property
 prop_unionWith =
   ttProp
@@ -450,12 +464,40 @@ prop_unionWith =
     M.unionWith
     NEM.unionWith
 
+prop_unionMapWithLeft :: Property
+prop_unionMapWithLeft =
+  ttProp
+    (gf2 valGen :?> GTMap :-> GTNEMap :-> TTNEMap)
+    M.unionWith
+    NEM.unionMapWithLeft
+
+prop_unionMapWithRight :: Property
+prop_unionMapWithRight =
+  ttProp
+    (gf2 valGen :?> GTNEMap :-> GTMap :-> TTNEMap)
+    M.unionWith
+    NEM.unionMapWithRight
+
 prop_unionWithKey :: Property
 prop_unionWithKey =
   ttProp
     (gf3 valGen :?> GTNEMap :-> GTNEMap :-> TTNEMap)
     M.unionWithKey
     NEM.unionWithKey
+
+prop_unionMapWithKeyLeft :: Property
+prop_unionMapWithKeyLeft =
+  ttProp
+    (gf3 valGen :?> GTMap :-> GTNEMap :-> TTNEMap)
+    M.unionWithKey
+    NEM.unionMapWithKeyLeft
+
+prop_unionMapWithKeyRight :: Property
+prop_unionMapWithKeyRight =
+  ttProp
+    (gf3 valGen :?> GTNEMap :-> GTMap :-> TTNEMap)
+    M.unionWithKey
+    NEM.unionMapWithKeyRight
 
 prop_unions :: Property
 prop_unions =