+++ CHANGELOG.md
@@ -7,6 +7,7 @@ Version 0.3.6.0
 *   Add `IsList` instances for `-XOverloadedLists`.
 *   Add `deleteMaybe` for `NEMap`, `NEIntMap`, `NESet`, and `NEIntSet`.
 *   Add indexed traversal instances for `NEMap`, `NEIntMap`, and `NESeq`.
+*   Add union functions for combining possibly-empty and non-empty maps and sets.
 *   Fix `NESeq.adjust'` to use the strict `Seq.adjust'` in the tail.
 
 Version 0.3.5.0